tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource.image.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.zerosource.image</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.zerosource.image</h1>
<p class="desc">drawImage with zero-sized source rectangle from image draws nothing without exception</p>


<script>
var t = async_test("drawImage with zero-sized source rectangle from image draws nothing without exception");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(async function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

await (function(canvas, ctx) {
ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
fetch('/images/red-zerowidth.svg')
  .then(response => response.blob())
    .then(blob => {
      createImageBitmap(blob)
        .then(bitmap => {
          ctx.drawImage(bitmap, 0, 0, 100, 50);
          ctx.drawImage(bitmap, 0, 0, 100, 50);
          ctx.drawImage(bitmap, 0, 0, 100, 50);
          _assertPixel(canvas, 50,25, 0,255,0,255);
      });
});
})(canvas, ctx);
t.done();

});
</script>
